+++ apollo-client/src/test/java/com/ctrip/framework/apollo/spring/JavaConfigAnnotationTest.java
@@ -41,6 +41,7 @@ import java.util.concurrent.TimeUnit;
 import org.junit.After;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
+import org.mockito.Mockito;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.springframework.beans.factory.BeanCreationException;
@@ -152,7 +153,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     String someValue = UUID.randomUUID().toString();
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, mock(Config.class));
     Config xxxConfig = mock(Config.class);
-    when(xxxConfig.getProperty(eq(someKey), anyString())).thenReturn(someValue);
+    when(xxxConfig.getProperty(eq(someKey), Mockito.nullable(String.class))).thenReturn(someValue);
     mockConfig("xxx", xxxConfig);
 
     TestEnableApolloConfigResolveExpressionWithDefaultValueConfiguration configuration =
@@ -160,7 +161,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
 
     // check
     assertEquals(someValue, configuration.getSomeKey());
-    verify(xxxConfig, times(1)).getProperty(eq(someKey), anyString());
+    verify(xxxConfig, times(1)).getProperty(eq(someKey), Mockito.nullable(String.class));
   }
 
   @Test
@@ -173,7 +174,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     System.setProperty(SystemPropertyKeyConstants.SIMPLE_NAMESPACE, resolvedNamespaceName);
 
     Config yyyConfig = mock(Config.class);
-    when(yyyConfig.getProperty(eq(someKey), anyString())).thenReturn(someValue);
+    when(yyyConfig.getProperty(eq(someKey), Mockito.nullable(String.class))).thenReturn(someValue);
     mockConfig(resolvedNamespaceName, yyyConfig);
 
     TestEnableApolloConfigResolveExpressionWithDefaultValueConfiguration configuration =
@@ -181,7 +182,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
 
     // check
     assertEquals(someValue, configuration.getSomeKey());
-    verify(yyyConfig, times(1)).getProperty(eq(someKey), anyString());
+    verify(yyyConfig, times(1)).getProperty(eq(someKey), Mockito.nullable(String.class));
   }
 
   @Test(expected = BeanCreationException.class)
@@ -210,7 +211,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     doAnswer(new Answer<Object>() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
-        applicationListeners.add(invocation.getArgumentAt(0, ConfigChangeListener.class));
+        applicationListeners.add(invocation.getArgument(0, ConfigChangeListener.class));
 
         return Void.class;
       }
@@ -219,7 +220,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     doAnswer(new Answer<Object>() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
-        fxApolloListeners.add(invocation.getArgumentAt(0, ConfigChangeListener.class));
+        fxApolloListeners.add(invocation.getArgument(0, ConfigChangeListener.class));
 
         return Void.class;
       }
@@ -283,7 +284,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     doAnswer(new Answer<Object>() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
-        applicationListeners.add(invocation.getArgumentAt(0, ConfigChangeListener.class));
+        applicationListeners.add(invocation.getArgument(0, ConfigChangeListener.class));
 
         return Void.class;
       }
@@ -292,7 +293,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     doAnswer(new Answer<Object>() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
-        fxApolloListeners.add(invocation.getArgumentAt(0, ConfigChangeListener.class));
+        fxApolloListeners.add(invocation.getArgument(0, ConfigChangeListener.class));
 
         return Void.class;
       }
@@ -341,10 +342,10 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     final ArgumentCaptor<Set> fxApolloConfigInterestedKeys = ArgumentCaptor.forClass(Set.class);
 
     verify(applicationConfig, times(2))
-        .addChangeListener(any(ConfigChangeListener.class), applicationConfigInterestedKeys.capture(), anySetOf(String.class));
+        .addChangeListener(any(ConfigChangeListener.class), applicationConfigInterestedKeys.capture(), Mockito.nullable(Set.class));
 
     verify(fxApolloConfig, times(1))
-        .addChangeListener(any(ConfigChangeListener.class), fxApolloConfigInterestedKeys.capture(), anySetOf(String.class));
+        .addChangeListener(any(ConfigChangeListener.class), fxApolloConfigInterestedKeys.capture(), Mockito.nullable(Set.class));
 
     assertEquals(2, applicationConfigInterestedKeys.getAllValues().size());
 
@@ -372,7 +373,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
         .forClass(Set.class);
 
     verify(applicationConfig, times(1))
-        .addChangeListener(any(ConfigChangeListener.class), anySetOf(String.class),
+        .addChangeListener(any(ConfigChangeListener.class), Mockito.nullable(Set.class),
             interestedKeyPrefixesArgumentCaptor.capture());
 
     assertEquals(1, interestedKeyPrefixesArgumentCaptor.getAllValues().size());
@@ -403,7 +404,7 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
         TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1.class, AppConfig11.class);
 
     verify(simpleConfig, atLeastOnce())
-        .addChangeListener(any(ConfigChangeListener.class), anySetOf(String.class),
+        .addChangeListener(any(ConfigChangeListener.class), Mockito.nullable(Set.class),
             anySetOf(String.class));
 
     Properties properties = new Properties();