+++ apollo-client/src/test/java/com/ctrip/framework/apollo/spring/JavaConfigPlaceholderTest.java
@@ -30,6 +30,7 @@ import com.ctrip.framework.apollo.spring.annotation.EnableApolloConfig;
 import java.util.List;
 import java.util.Properties;
 import org.junit.Test;
+import org.mockito.Mockito;
 import org.springframework.beans.factory.BeanCreationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -59,8 +60,8 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(config.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(config.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(config.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -80,8 +81,8 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(config.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(config.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(config.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -126,11 +127,11 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config application = mock(Config.class);
-    when(application.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
+    when(application.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, application);
 
     Config fxApollo = mock(Config.class);
-    when(application.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(application.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
     mockConfig(FX_APOLLO_NAMESPACE, fxApollo);
 
     check(someTimeout, someBatch, AppConfig3.class);
@@ -152,7 +153,7 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     mockConfigFile("application.yml", configFile);
 
     Config fxApollo = mock(Config.class);
-    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(anotherTimeout));
+    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(anotherTimeout));
     mockConfig(FX_APOLLO_NAMESPACE, fxApollo);
 
     check(someTimeout, someBatch, AppConfig11.class);
@@ -166,12 +167,12 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config fxApollo = mock(Config.class);
-    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(fxApollo.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(fxApollo.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
     mockConfig(FX_APOLLO_NAMESPACE, fxApollo);
 
     Config application = mock(Config.class);
-    when(application.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(anotherTimeout));
+    when(application.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(anotherTimeout));
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, application);
 
     check(someTimeout, someBatch, AppConfig6.class);
@@ -185,12 +186,12 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config fxApollo = mock(Config.class);
-    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(fxApollo.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(fxApollo.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
     mockConfig(FX_APOLLO_NAMESPACE, fxApollo);
 
     Config application = mock(Config.class);
-    when(application.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(anotherTimeout));
+    when(application.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(anotherTimeout));
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, application);
 
     check(someTimeout, someBatch, AppConfig6.class);
@@ -203,12 +204,12 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config application = mock(Config.class);
-    when(application.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(application.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(application.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(application.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, application);
 
     Config fxApollo = mock(Config.class);
-    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(anotherTimeout));
+    when(fxApollo.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(anotherTimeout));
     mockConfig(FX_APOLLO_NAMESPACE, fxApollo);
 
     check(anotherTimeout, someBatch, AppConfig2.class, AppConfig4.class);
@@ -220,8 +221,8 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(config.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(config.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(config.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -239,8 +240,8 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someBatch = 2000;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(TIMEOUT_PROPERTY), anyString())).thenReturn(String.valueOf(someTimeout));
-    when(config.getProperty(eq(BATCH_PROPERTY), anyString())).thenReturn(String.valueOf(someBatch));
+    when(config.getProperty(eq(TIMEOUT_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someTimeout));
+    when(config.getProperty(eq(BATCH_PROPERTY), Mockito.nullable(String.class))).thenReturn(String.valueOf(someBatch));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -259,9 +260,9 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someValue = 1234;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(a), anyString())).thenReturn(a);
-    when(config.getProperty(eq(b), anyString())).thenReturn(b);
-    when(config.getProperty(eq(String.format("%s.%s", a, b)), anyString()))
+    when(config.getProperty(eq(a), Mockito.nullable(String.class))).thenReturn(a);
+    when(config.getProperty(eq(b), Mockito.nullable(String.class))).thenReturn(b);
+    when(config.getProperty(eq(String.format("%s.%s", a, b)), Mockito.nullable(String.class)))
         .thenReturn(String.valueOf(someValue));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
@@ -281,9 +282,9 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someValue = 1234;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(a), anyString())).thenReturn(a);
-    when(config.getProperty(eq(b), anyString())).thenReturn(b);
-    when(config.getProperty(eq(c), anyString())).thenReturn(String.valueOf(someValue));
+    when(config.getProperty(eq(a), Mockito.nullable(String.class))).thenReturn(a);
+    when(config.getProperty(eq(b), Mockito.nullable(String.class))).thenReturn(b);
+    when(config.getProperty(eq(c), Mockito.nullable(String.class))).thenReturn(String.valueOf(someValue));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -300,8 +301,8 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     String b = "b";
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(a), anyString())).thenReturn(a);
-    when(config.getProperty(eq(b), anyString())).thenReturn(b);
+    when(config.getProperty(eq(a), Mockito.nullable(String.class))).thenReturn(a);
+    when(config.getProperty(eq(b), Mockito.nullable(String.class))).thenReturn(b);
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -321,10 +322,10 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     int someValue = 1234;
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(a), anyString())).thenReturn(a);
-    when(config.getProperty(eq(b), anyString())).thenReturn(b);
-    when(config.getProperty(eq(String.format("%s.%s", a, b)), anyString())).thenReturn(nestedProperty);
-    when(config.getProperty(eq(nestedKey), anyString())).thenReturn(String.valueOf(someValue));
+    when(config.getProperty(eq(a), Mockito.nullable(String.class))).thenReturn(a);
+    when(config.getProperty(eq(b), Mockito.nullable(String.class))).thenReturn(b);
+    when(config.getProperty(eq(String.format("%s.%s", a, b)), Mockito.nullable(String.class))).thenReturn(nestedProperty);
+    when(config.getProperty(eq(nestedKey), Mockito.nullable(String.class))).thenReturn(String.valueOf(someValue));
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -344,9 +345,9 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     String nestedProperty = String.format("${%s:%d}", nestedKey, someValue);
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(a), anyString())).thenReturn(a);
-    when(config.getProperty(eq(b), anyString())).thenReturn(b);
-    when(config.getProperty(eq(String.format("%s.%s", a, b)), anyString())).thenReturn(nestedProperty);
+    when(config.getProperty(eq(a), Mockito.nullable(String.class))).thenReturn(a);
+    when(config.getProperty(eq(b), Mockito.nullable(String.class))).thenReturn(b);
+    when(config.getProperty(eq(String.format("%s.%s", a, b)), Mockito.nullable(String.class))).thenReturn(nestedProperty);
 
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
@@ -363,9 +364,9 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     String otherJson = "[{\"a\":\"otherString\", \"b\":10},{\"a\":\"astring2\", \"b\":20}]";
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(JSON_PROPERTY), anyString())).thenReturn(someJson);
-    when(config.getProperty(eq(OTHER_JSON_PROPERTY), anyString())).thenReturn(otherJson);
-    when(config.getProperty(eq("a"), anyString())).thenReturn(JSON_PROPERTY);
+    when(config.getProperty(eq(JSON_PROPERTY), Mockito.nullable(String.class))).thenReturn(someJson);
+    when(config.getProperty(eq(OTHER_JSON_PROPERTY), Mockito.nullable(String.class))).thenReturn(otherJson);
+    when(config.getProperty(eq("a"), Mockito.nullable(String.class))).thenReturn(JSON_PROPERTY);
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
     AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(
@@ -389,9 +390,9 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     String someInvalidJson = "someInvalidJson";
 
     Config config = mock(Config.class);
-    when(config.getProperty(eq(JSON_PROPERTY), anyString())).thenReturn(someInvalidJson);
-    when(config.getProperty(eq(OTHER_JSON_PROPERTY), anyString())).thenReturn(someInvalidJson);
-    when(config.getProperty(eq("a"), anyString())).thenReturn(JSON_PROPERTY);
+    when(config.getProperty(eq(JSON_PROPERTY), Mockito.nullable(String.class))).thenReturn(someInvalidJson);
+    when(config.getProperty(eq(OTHER_JSON_PROPERTY), Mockito.nullable(String.class))).thenReturn(someInvalidJson);
+    when(config.getProperty(eq("a"), Mockito.nullable(String.class))).thenReturn(JSON_PROPERTY);
     mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
 
     new AnnotationConfigApplicationContext(AppConfig8.class).getBean(TestJsonPropertyBean.class);