+++ apollo-client/src/test/java/com/ctrip/framework/apollo/spring/config/ConfigPropertySourceTest.java
@@ -99,7 +99,7 @@ public class ConfigPropertySourceTest {
     doAnswer(new Answer() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
-        listeners.add(invocation.getArgumentAt(0, ConfigChangeListener.class));
+        listeners.add(invocation.getArgument(0, ConfigChangeListener.class));
 
         return Void.class;
       }